Fix CC issues
